Final modifications on the automatic layout structure.
